requirements.txt
requests>=2.31.0
psutil>=5.9.0
rich>=13.0.0
pywin32>=306
pyautogui>=0.9.54
cryptography>=41.0.0
playwright>=1.40.0
pynput>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
customtkinter>=5.2.0
Pillow>=10.0.0
